generate_m3u/requirements.txt
mysql-connector-python==9.0.0
tqdm==4.66.5
